tests/requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Data handling
pydantic>=1.8.0
pyyaml>=5.4.0

# ML and AI
openai>=1.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# Networking and APIs
requests>=2.25.0
websocket-client>=1.0.0

# Simulation and aerospace
bluesky-simulator>=1.0.0  # BlueSky ATM simulator

# CLI and utilities
click>=8.0.0
tqdm>=4.60.0
rich>=10.0.0

# Development dependencies
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910
pre-commit>=2.13.0
